requirements.txt
selenium
requests
Pillow
opencv-python
numpy
webdriver_manager